Add factoring out components and translations
