+++ packages/common/src/hooks/useOpenInNewTab/useOpenInNewTab.stories.tsx
@@ -1,7 +1,7 @@
 import React, { useState } from 'react';
 import { Story } from '@storybook/react';
 import { useOpenInNewTab } from './useOpenInNewTab';
-import { BaseButton, BasicTextInput } from '@common/components';
+import { Typography, BaseButton, BasicTextInput } from '@common/components';
 
 export default {
   title: 'Hooks/useOpenInNewTab',
@@ -13,6 +13,7 @@ const Template: Story = () => {
 
   return (
     <div>
+      <Typography>Enter a relative URL</Typography>
       <BasicTextInput value={url} onChange={e => setUrl(e.target.value)} />
       <BaseButton onClick={() => openInNewTab(url)}>Click Me</BaseButton>
     </div>
