+++ packages/common/src/intl/locales/en/distribution.json
@@ -1,32 +1,29 @@
 {
+  "button.add-from-master-list": "Add from master list",
+  "button.add-item": "Add item",
+  "button.create-shipment": "Create Shipment",
   "button.delete-lines": "Delete selected lines",
   "button.new-outbound-shipment": "New Outbound Shipment",
   "button.new-shipment": "New Shipment",
-  "button.create-shipment": "Create Shipment",
   "button.save-and-confirm-status": "Confirm {{status}}",
   "button.select-a-color": "Select a color",
-  "button.add-from-master-list": "Add from master list",
-  "button.add-item": "Add item",
-  "dev.log-draft": "Log draft",
+  "error.failed-to-create-outbound": "Failed to create Outbound Shipment",
   "error.requisition-not-found": "Requisition not found",
   "error.shipment-not-found": "Shipment not found",
-  "warning.nothing-to-supply": "Nothing left to supply!",
-  "error.failed-to-create-outbound": "Failed to create Outbound Shipment",
-  "heading.comment": "Comment",
+  "heading.are-you-sure": "Are you sure?",
   "heading.cannot-do-that": "Can't do that!",
+  "heading.charges": "Charges",
+  "heading.comment": "Comment",
+  "heading.create-outbound-shipment": "Create Outbound Shipment",
   "heading.edit-item": "Edit Item",
   "heading.order": "Order",
   "heading.related-documents": "Related documents",
-  "heading.charges": "Charges",
-  "heading.stock-charges": "Stock Charges",
   "heading.service-charges": "Service Charges",
-  "heading.totals": "Totals",
+  "heading.stock-charges": "Stock Charges",
+  "heading.stock-details": "Stock details",
   "heading.sub-total": "Sub total",
   "heading.total": "Total",
-  "heading.stock-details": "Stock details",
-  "heading.are-you-sure": "Are you sure?",
-  "message.confirm-add-from-master-list": "Do you want to add all of the items from this master list?",
-  "heading.create-outbound-shipment": "Create Outbound Shipment",
+  "heading.totals": "Totals",
   "label.add-batch": "Add batch",
   "label.allocated": "Allocated",
   "label.allocation": "Allocation",
@@ -49,16 +46,18 @@
   "label.units-in-pack-size-of_one": "unit in pack size of",
   "label.units-in-pack-size-of_other": "units in pack size of",
   "messages.cant-delete-invoices": "Can only delete invoices with a status of 'New' or 'Allocated'",
-  "messages.must-allocate-all-lines": "Cannot change the status until all lines have been allocated.",
   "messages.cant-delete-requisitions": "Can only delete requisitions with a status of 'Draft'",
   "messages.click-to-return-to-requisitions": "Unable to find a requisition with that ID. Click OK to return to the requisition list",
   "messages.click-to-return-to-shipments": "Unable to find a shipment with that ID. Click OK to return to the shipment list",
+  "messages.confirm-add-from-master-list": "Do you want to add all of the items from this master list?",
   "messages.confirm-status-as": "Confirm status as $t({{status}})?",
+  "messages.create-outbound-from-requisition": "Are you sure you want to create an Outbound Shipment from this Customer Requisition?",
   "messages.deleted-invoices": "Deleted {{number}} invoices",
   "messages.deleted-lines": "Deleted {{number}} lines",
   "messages.deleted-requisitions": "Deleted {{number}} requisitions",
   "messages.error-saving-shipment": "Error saving shipment 🥺",
+  "messages.must-allocate-all-lines": "Cannot change the status until all lines have been allocated.",
   "messages.select-rows-to-delete": "Select rows to delete",
   "messages.shipment-saved": "Shipment saved 🥳",
-  "messages.create-outbound-from-requisition": "Are you sure you want to create an Outbound Shipment from this Customer Requisition?"
+  "warning.nothing-to-supply": "Nothing left to supply!"
 }
