+++ packages/common/src/intl/locales/en/replenishment.json
@@ -2,35 +2,37 @@
   "button.delete-lines": "Delete lines",
   "button.new-shipment": "New Shipment",
   "button.save-and-confirm-status": "Confirm {{status}}",
-  "dev.log-draft": "Log draft",
   "error.requisition-not-found": "Requisition not found",
   "error.shipment-not-found": "Shipment not found",
+  "heading.are-you-sure": "Are you sure?",
   "heading.related-documents": "Related documents",
   "label.add-batch": "Add batch",
+  "label.draft": "Draft",
+  "label.finalised": "Finalised",
   "label.group-by-item": "Group by Item",
   "label.location": "Location",
+  "label.new": "New",
+  "label.order-quantity": "Order quantity",
   "label.pricing": "Pricing",
   "label.quantities": "Quantities",
   "label.select-rows-to-delete-them": "Select rows to delete them",
+  "label.sent": "Sent",
+  "label.stock-details": "Stock details",
+  "label.suggested-quantity": "Suggested quantity",
   "label.supplier-name": "Supplier Name",
   "label.supplier-ref": "Supplier Ref",
+  "label.min-months-of-stock": "Min. Months of Stock",
+  "label.number-months": "{{numberOfMonths}} months",
   "messages.cant-delete-invoices": "Can only delete invoices with a status of 'New'",
   "messages.cant-delete-requisitions": "Can only delete requisitions with a status of 'Draft'",
-  "messages.deleted-lines": "Deleted {{number}} lines",
-  "messages.deleted-requisitions": "Deleted {{number}} requisitions",
-  "messages.select-rows-to-delete": "Select rows to delete",
-  "messages.deleted-invoices": "Deleted {{number}} invoices",
+  "messages.changing-min-mos": "This will change the minimum months of stock threshold.",
   "messages.click-to-return-to-requisitions": "Unable to find a requisition with that ID. Click OK to return to the requisition list",
   "messages.click-to-return-to-shipments": "Unable to find a shipment with that ID. Click OK to return to the shipment list",
-  "label.draft": "Draft",
-  "label.sent": "Sent",
-  "label.new": "New",
-  "label.finalised": "Finalised",
-  "heading.are-you-sure": "Are you sure?",
-  "messages.saved": "Saved",
   "messages.confirm-status-as": "Confirm status as $t({{status}})?",
   "messages.could-not-save": "Could not Save",
-  "label.stock-details": "Stock details",
-  "label.order-quantity": "Order quantity",
-  "label.suggested-quantity": "Suggested quantity"
+  "messages.deleted-invoices": "Deleted {{number}} invoices",
+  "messages.deleted-lines": "Deleted {{number}} lines",
+  "messages.deleted-requisitions": "Deleted {{number}} requisitions",
+  "messages.saved": "Saved",
+  "messages.select-rows-to-delete": "Select rows to delete"
 }
